fixed AttachToConsole which was failing on Windows XP
On windows xp, the first call to AttachConsole was always failing with
error 0x6 (The handle is invalid).
Subsequent calls were failing too with error 0x5 (Access is denied).
Sleeping inside a loop after the failed attempt (even upto 5000ms)
didn't help.

On windows 7 the first call was failing as well but with error 0x1f (A
device attached to the system is not functioning).
Yet after a short sleep (even as low as 10ms) the second call usually
succeeded.

BUT, when the sleep was called BEFORE calling AttachConsole, on windows
xp the call succeeded in the first attempt!
On windows 7 the success was in the first try as well.